Create Fun Photo Filters
This user-friendly program allows users to add filters to their photos in real-time. The program will ask the user to enter the file name of the image they'd wish to jazz up. Once the image is fully uploaded, users will be able to add and adjust colors of images to their liking. Once user is finished with their filtered photo, they will be given the option to save or discard.