Aerial Imaging
I made this code for my aerial imaging project in my engineering class. The purpose of the project was to create a system that could deliver objects from point A to B. We used a drone to lift our craft, Arduino software, an altimeter, a servo motor, many wires, and a camera to bring this project to life. My teammates and I built a structure for the items to be protected during the craft's departure and landing. The program records photographs from its highest altitude and of the item once its delivered to the target destination.